feat: add parallelism and concurrency to image processing
Also cleared out GDI+ dead code. Flip and transpose mimics a 90 degree clockwise rotation and is computationally less expensive as it's fewer matrix multiplication operations that a standard rotation and achieves same effect.

No fiddling directly with GPU as that is too specific to certain machines, spec was for support low end AND high end PCs so this is not a guaranteed win. Make use of a decent core which most machines have nowadays.